changes for Telemetry Data Type Update and Frontend Monitoring

+++ backend/src/main/java/com/productivityx/dto/TelemetryBatchRequest.java
@@ -8,7 +8,7 @@ import java.util.Map;
 public class TelemetryBatchRequest {
     private String eventId;
     private OffsetDateTime timestamp;
-    private Integer focusScore;
+    private Double focusScore;
     private Integer awaySeconds;
     private Integer idleSeconds;
     private String cameraState;
